requirements.txt
Django==3.2
django-cors-headers==3.7.0
channels==3.0.3
channels-redis==3.2.0
Pillow==8.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
mysqlclient==2.0.3
psycopg2==2.8.6
xlrd==2.0.1
pandas==1.2.4
openpyxl===3.0.7
gunicorn===20.1.0